refactor everything into word view
